crop-price-backend/requirements.txt
fastapi
uvicorn
pandas
geopy
python-dotenv
requests
